+++ docker-compose.yml
@@ -2,7 +2,7 @@ version: '3.2'
 services:
   app:
     build:
-      context: app_
+      context: app
     ports:
       - '8000:8000'
     depends_on:
